fix: Edit project structure as using central server
