+++ scripts/install.sh
@@ -3,62 +3,50 @@
 # 필요한 패키지 설치
 sudo apt-get update
 sudo apt-get install -y python3 python3-pip
-sudo pip3 install docker requests flask
+sudo pip3 install docker requests
 
 # 디렉토리 생성
-sudo mkdir -p /opt/company-security-solution/{bin,lib,config}
-sudo mkdir -p /var/log/company-security-solution
+sudo mkdir -p /opt/fish-hash-bowl/{src,config}
+sudo mkdir -p /var/log/fish-hash-bowl
 
 # 파일 복사
-sudo cp bin/docker-wrapper /opt/company-security-solution/bin/
-sudo cp lib/integrity_check.py /opt/company-security-solution/lib/
-sudo cp lib/register_image_hash.py /opt/company-security-solution/lib/
-sudo cp docker-security-plugin.py /opt/company-security-solution/
+sudo cp src/image_utils.py /opt/fish-hash-bowl/src/
+sudo cp src/integrity_checker.py /opt/fish-hash-bowl/src/
+sudo cp src/hash_registrar.py /opt/fish-hash-bowl/src/
+sudo cp src/event_monitor.py /opt/fish-hash-bowl/src/
 
 # 권한 설정
-sudo chmod +x /opt/company-security-solution/bin/docker-wrapper
-sudo chmod +x /opt/company-security-solution/lib/integrity_check.py
-sudo chmod +x /opt/company-security-solution/lib/register_image_hash.py
-sudo chmod +x /opt/company-security-solution/docker-security-plugin.py
-
-# Docker 명령어 대체
-sudo mv /usr/bin/docker /usr/bin/docker.original
-sudo ln -s /opt/company-security-solution/bin/docker-wrapper /usr/bin/docker
+sudo chmod +x /opt/fish-hash-bowl/src/integrity_checker.py
+sudo chmod +x /opt/fish-hash-bowl/src/hash_registrar.py
+sudo chmod +x /opt/fish-hash-bowl/src/event_monitor.py
 
 # 로그 디렉토리 권한 설정
-sudo chown -R root:root /var/log/company-security-solution
-sudo chmod 755 /var/log/company-security-solution
+sudo chown -R root:root /var/log/fish-hash-bowl
+sudo chmod 755 /var/log/fish-hash-bowl
 
 # API 키 설정
 echo "Enter your API key:"
 read API_KEY
 
 # config.json 생성
-cat > /opt/company-security-solution/config/config.json << EOF
+cat > /opt/fish-hash-bowl/config/config.json << EOF
 {
-    "api_endpoint": "https://www.check-docker.com/api/v1",
+    "central_server_url": "https://www.check-docker.com/api/v1",
     "api_key": "$API_KEY"
 }
 EOF
 
 # config.json 권한 설정
-sudo chmod 600 /opt/company-security-solution/config/config.json
+sudo chmod 600 /opt/fish-hash-bowl/config/config.json
 
-# Docker 데몬 설정 수정
-sudo tee /etc/docker/daemon.json > /dev/null <<EOF
-{
-  "authorization-plugins": ["docker-security-plugin"]
-}
-EOF
-
-# Docker 보안 플러그인 서비스 파일 생성
-sudo tee /etc/systemd/system/docker-security-plugin.service > /dev/null <<EOF
+# Docker 이벤트 모니터 서비스 파일 생성
+sudo tee /etc/systemd/system/docker-event-monitor.service > /dev/null <<EOF
 [Unit]
-Description=Docker Security Plugin
+Description=Docker Event Monitor
 After=docker.service
 
 [Service]
-ExecStart=/usr/bin/python3 /opt/company-security-solution/docker-security-plugin.py
+ExecStart=/usr/bin/python3 /opt/fish-hash-bowl/src/event_monitor.py
 Restart=always
 
 [Install]
@@ -67,10 +55,16 @@ EOF
 
 # 서비스 활성화 및 시작
 sudo systemctl daemon-reload
-sudo systemctl enable docker-security-plugin.service
-sudo systemctl start docker-security-plugin.service
-
-# Docker 재시작
-sudo systemctl restart docker
+sudo systemctl enable docker-event-monitor.service
+sudo systemctl start docker-event-monitor.service
 
 echo "Company security solution installed successfully."
+
+# TODO:
+# 1. 설치 과정 로깅 추가
+# 2. 설치 실패 시 롤백 메커니즘 구현
+# 3. 사용자 입력 검증 (예: API 키 형식 확인)
+# 4. 의존성 검사 (예: Docker 설치 여부, Python 버전 확인)
+# 5. 설치 완료 후 시스템 상태 확인 및 보고
+# 6. 업그레이드 메커니즘 구현
+# 7. 설치 옵션 추가 (예: 사용자 정의 설치 경로, 로그 레벨 설정 등)
